# openapi.yaml
openapi: 3.0.3
info:
  title: Cari Importer API
  description: Excel dosyalarından cari hesap verilerini import etmek için API
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /imports/excel:
    post:
      summary: Excel dosyası yükle ve import et
      description: Excel dosyasını yükleyip cari hesap verilerini import eder
      tags:
        - Imports
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Yüklenecek Excel dosyası
              required:
                - file
      responses:
        '201':
          description: Dosya başarıyla yüklendi ve işlendi
          content:
            application/json:
              schema:
                type: object
                properties:
                  file:
                    $ref: '#/components/schemas/StagingFile'
                  report:
                    $ref: '#/components/schemas/ImportReport'
        '400':
          description: Dosya eksik veya geçersiz
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Sunucu hatası
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /imports/{id}:
    get:
      summary: Import durumunu kontrol et
      description: Belirtilen ID'ye sahip import işleminin durumunu getirir
      tags:
        - Imports
      parameters:
        - name: id
          in: path
          required: true
          description: Import ID'si
          schema:
            type: integer
      responses:
        '200':
          description: Import durumu başarıyla getirildi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StagingFile'
        '404':
          description: Import bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/{code}/balances:
    get:
      summary: Müşteri bakiye geçmişi
      description: Belirtilen müşteri koduna sahip müşterinin bakiye geçmişini getirir
      tags:
        - Customers
      parameters:
        - name: code
          in: path
          required: true
          description: Müşteri kodu
          schema:
            type: string
      responses:
        '200':
          description: Müşteri bakiye geçmişi başarıyla getirildi
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    $ref: '#/components/schemas/Customer'
                  snapshots:
                    type: array
                    items:
                      $ref: '#/components/schemas/CustomerBalanceSnapshot'
        '404':
          description: Müşteri bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /customers/{code}/transactions:
    get:
      summary: Müşteri işlemleri
      description: Belirtilen müşteri koduna sahip müşterinin işlemlerini getirir
      tags:
        - Customers
      parameters:
        - name: code
          in: path
          required: true
          description: Müşteri kodu
          schema:
            type: string
        - name: page
          in: query
          description: Sayfa numarası
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          description: Sayfa başına kayıt sayısı
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Müşteri işlemleri başarıyla getirildi
          content:
            application/json:
              schema:
                type: object
                properties:
                  customer:
                    $ref: '#/components/schemas/Customer'
                  transactions:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
                  pagination:
                    type: object
                    properties:
                      page:
                        type: integer
                      limit:
                        type: integer
                      total:
                        type: integer
                      totalPages:
                        type: integer
                      hasNext:
                        type: boolean
                      hasPrev:
                        type: boolean
        '404':
          description: Müşteri bulunamadı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /transactions:
    get:
      summary: İşlem listesi
      description: Filtrelenebilir işlem listesi getirir
      tags:
        - Transactions
      parameters:
        - name: code
          in: query
          description: Müşteri kodu
          schema:
            type: string
        - name: from
          in: query
          description: Başlangıç tarihi (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Bitiş tarihi (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: type
          in: query
          description: Belge türü
          schema:
            type: string
        - name: voucher
          in: query
          description: Evrak numarası (içerir)
          schema:
            type: string
        - name: min
          in: query
          description: Minimum tutar (TL)
          schema:
            type: number
        - name: max
          in: query
          description: Maksimum tutar (TL)
          schema:
            type: number
        - name: page
          in: query
          description: Sayfa numarası
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Sayfa başına kayıt sayısı
          schema:
            type: integer
            default: 50
        - name: sort
          in: query
          description: Sıralama
          schema:
            type: string
            enum: [date_asc, date_desc, amount_asc, amount_desc]
            default: date_desc
      responses:
        '200':
          description: İşlem listesi başarıyla getirildi
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Toplam kayıt sayısı
                  page:
                    type: integer
                    description: Mevcut sayfa
                  pageSize:
                    type: integer
                    description: Sayfa başına kayıt sayısı
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
  /transactions/by-customer/{code}:
    get:
      summary: Müşteri işlemleri (alternatif endpoint)
      description: Belirtilen müşteri koduna sahip müşterinin işlemlerini getirir (filtreleme destekli)
      tags:
        - Transactions
      parameters:
        - name: code
          in: path
          required: true
          description: Müşteri kodu
          schema:
            type: string
        - name: from
          in: query
          description: Başlangıç tarihi (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: to
          in: query
          description: Bitiş tarihi (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: type
          in: query
          description: Belge türü
          schema:
            type: string
        - name: voucher
          in: query
          description: Evrak numarası (içerir)
          schema:
            type: string
        - name: min
          in: query
          description: Minimum tutar (TL)
          schema:
            type: number
        - name: max
          in: query
          description: Maksimum tutar (TL)
          schema:
            type: number
        - name: page
          in: query
          description: Sayfa numarası
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          description: Sayfa başına kayıt sayısı
          schema:
            type: integer
            default: 50
        - name: sort
          in: query
          description: Sıralama
          schema:
            type: string
            enum: [date_asc, date_desc, amount_asc, amount_desc]
            default: date_desc
      responses:
        '200':
          description: Müşteri işlemleri başarıyla getirildi
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    description: Toplam kayıt sayısı
                  page:
                    type: integer
                    description: Mevcut sayfa
                  pageSize:
                    type: integer
                    description: Sayfa başına kayıt sayısı
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Transaction'
...
components:
  schemas:
    StagingFile:
      type: object
      properties:
        id:
          type: integer
          description: Dosya ID'si
        originalFilename:
          type: string
          description: Orijinal dosya adı
        uploadedAt:
          type: string
          format: date-time
          description: Yükleme tarihi
        status:
          type: string
          enum: [PENDING, PROCESSING, PROCESSED, FAILED]
          description: İşlem durumu
        rowCount:
          type: integer
          description: Toplam satır sayısı
        insertedCount:
          type: integer
          description: Eklenen kayıt sayısı
        updatedCount:
          type: integer
          description: Güncellenen kayıt sayısı
        skippedCount:
          type: integer
          description: Atlanan kayıt sayısı
        error:
          type: string
          nullable: true
          description: Hata mesajı (varsa)
        customerBalances:
          type: array
          items:
            $ref: '#/components/schemas/CustomerBalanceSnapshot'
    ImportReport:
      type: object
      properties:
        total:
          type: integer
          description: Toplam işlenen kayıt sayısı
        inserted:
          type: integer
          description: Eklenen kayıt sayısı
        updated:
          type: integer
          description: Güncellenen kayıt sayısı
        skipped:
          type: integer
          description: Atlanan kayıt sayısı
    Customer:
      type: object
      properties:
        id:
          type: integer
          description: Müşteri ID'si
        code:
          type: string
          description: Müşteri kodu
        name:
          type: string
          description: Müşteri adı
        phone:
          type: string
          nullable: true
          description: Telefon numarası
        address:
          type: string
          nullable: true
          description: Adres
        accountType:
          type: string
          nullable: true
          description: Hesap türü
        tag1:
          type: string
          nullable: true
          description: Özel kod 1
        tag2:
          type: string
          nullable: true
          description: Özel kod 2
    Transaction:
      type: object
      properties:
        id:
          type: integer
          description: İşlem ID'si
        customer:
          $ref: '#/components/schemas/Customer'
        docType:
          type: string
          nullable: true
          description: Belge türü
        txnDate:
          type: string
          format: date-time
          description: İşlem tarihi
        voucherNo:
          type: string
          nullable: true
          description: Evrak numarası
        description:
          type: string
          nullable: true
          description: Açıklama
        dueDate:
          type: string
          format: date-time
          nullable: true
          description: Vade tarihi
        amountBase:
          type: number
          format: float
          nullable: true
          description: Matrah tutarı (TL)
        discount:
          type: number
          format: float
          nullable: true
          description: İndirim tutarı (TL)
        amountNet:
          type: number
          format: float
          nullable: true
          description: Net tutar (TL)
        vat:
          type: number
          format: float
          nullable: true
          description: KDV tutarı (TL)
        debit:
          type: number
          format: float
          description: Borç tutarı (TL)
        credit:
          type: number
          format: float
          description: Alacak tutarı (TL)
        currency:
          type: string
          description: Para birimi
    CustomerBalanceSnapshot:
      type: object
      properties:
        id:
          type: integer
          description: Snapshot ID'si
        customerId:
          type: integer
          description: Müşteri ID'si
        fileId:
          type: integer
          description: Dosya ID'si
        reportedTotalDebit:
          type: number
          format: float
          description: Raporlanan toplam borç
        reportedTotalCredit:
          type: number
          format: float
          description: Raporlanan toplam alacak
        reportedDebtBalance:
          type: number
          format: float
          description: Raporlanan borç bakiyesi
        reportedCreditBalance:
          type: number
          format: float
          description: Raporlanan alacak bakiyesi
        calcTotalDebit:
          type: number
          format: float
          description: Hesaplanan toplam borç
        calcTotalCredit:
          type: number
          format: float
          description: Hesaplanan toplam alacak
        calcDebtBalance:
          type: number
          format: float
          description: Hesaplanan borç bakiyesi
        calcCreditBalance:
          type: number
          format: float
          description: Hesaplanan alacak bakiyesi
        diffTotalDebit:
          type: number
          format: float
          description: Toplam borç farkı
        diffTotalCredit:
          type: number
          format: float
          description: Toplam alacak farkı
        diffDebtBalance:
          type: number
          format: float
          description: Borç bakiyesi farkı
        diffCreditBalance:
          type: number
          format: float
          description: Alacak bakiyesi farkı
        createdAt:
          type: string
          format: date-time
          description: Oluşturulma tarihi
        file:
          $ref: '#/components/schemas/StagingFile'
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Hata mesajı
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...